requirements.txt
attrs==20.3.0
aws-cdk.assets==1.114.0
aws-cdk.aws-apigateway==1.114.0
aws-cdk.aws-applicationautoscaling==1.114.0
aws-cdk.aws-autoscaling==1.114.0
aws-cdk.aws-autoscaling-common==1.114.0
aws-cdk.aws-autoscaling-hooktargets==1.114.0
aws-cdk.aws-certificatemanager==1.114.0
aws-cdk.aws-cloudformation==1.114.0
aws-cdk.aws-cloudfront==1.114.0
aws-cdk.aws-cloudwatch==1.114.0
aws-cdk.aws-codebuild==1.114.0
aws-cdk.aws-codecommit==1.114.0
aws-cdk.aws-codeguruprofiler==1.114.0
aws-cdk.aws-codepipeline==1.114.0
aws-cdk.aws-codestarnotifications==1.114.0
aws-cdk.aws-cognito==1.114.0
aws-cdk.aws-ec2==1.114.0
aws-cdk.aws-ecr==1.114.0
aws-cdk.aws-ecr-assets==1.114.0
aws-cdk.aws-ecs==1.114.0
aws-cdk.aws-ecs-patterns==1.114.0
aws-cdk.aws-efs==1.114.0
aws-cdk.aws-elasticloadbalancing==1.114.0
aws-cdk.aws-elasticloadbalancingv2==1.114.0
aws-cdk.aws-events==1.114.0
aws-cdk.aws-events-targets==1.114.0
aws-cdk.aws-globalaccelerator==1.114.0
aws-cdk.aws-iam==1.114.0
aws-cdk.aws-kinesis==1.114.0
aws-cdk.aws-kinesisfirehose==1.114.0
aws-cdk.aws-kms==1.114.0
aws-cdk.aws-lambda==1.114.0
aws-cdk.aws-logs==1.114.0
aws-cdk.aws-route53==1.114.0
aws-cdk.aws-route53-targets==1.114.0
aws-cdk.aws-s3==1.114.0
aws-cdk.aws-s3-assets==1.114.0
aws-cdk.aws-sam==1.114.0
aws-cdk.aws-secretsmanager==1.114.0
aws-cdk.aws-servicediscovery==1.114.0
aws-cdk.aws-signer==1.114.0
aws-cdk.aws-sns==1.114.0
aws-cdk.aws-sns-subscriptions==1.114.0
aws-cdk.aws-sqs==1.114.0
aws-cdk.aws-ssm==1.114.0
aws-cdk.aws-stepfunctions==1.114.0
aws-cdk.cloud-assembly-schema==1.114.0
aws-cdk.core==1.114.0
aws-cdk.custom-resources==1.114.0
aws-cdk.cx-api==1.114.0
aws-cdk.region-info==1.114.0
cattrs==1.5.0
constructs==3.3.97
jsii==1.31.0
# Editable install with no version control (no-more-ransom==0.0.1)
-e .
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typing-extensions==3.10.0.0
